Rearrange test files
