+++ test/read/resolve-resource.test.js
